Revert "cambio en las migraciones para poner las foreign keys como INT y no BIGINT"
This reverts commit fa1107feff3d878676d12db1670504d208fbe667.

+++ database/migrations/2024_01_04_163106_create_usuarios_table.php
@@ -13,7 +13,7 @@ return new class extends Migration {
             $table->id();
             $table->string('dni', 9)->unique();
             $table->string('nombre');
-            $table->unsignedInteger('genero_id');
+            $table->unsignedBigInteger('genero_id');
             $table->foreign('genero_id')->references('id')->on('generos_tabla_maestra');
             $table->string('fecha_nacimiento');
             $table->string('email')->unique();
